refactor : 유저 엔티티 패키지명 변경 (#21)
