+++ entity/user/user.go
@@ -1,4 +1,4 @@
-package entity
+package user
 
 import "guestbook/entity"
